p097: slightly cheating solution
this only computes the last 9 digits
because the modulus in powmod is limited to half the width
of the number to be exponentiated, which is at most 64/2 bits with
built-ins. 10^10 does not fit that. 10^9 does, however.

but only 9 possibilities remain and those can be trial and errored on
PE's site